remove redundant code, fix bugs in lr.step, merge GoodsDataset into Vehicle
